app.py /requirements.txt
requests
openai
python-dotenv
beautifulsoup4

PyMuPDF
youtube-search-python
streamlit
selenium
